Add '-progress' option
Pursuant to https://github.com/pauldreik/rdfind/issues/30, this commit
adds a '-progress' option. If set to true, then a function pointer is
passed to 'fillwithbytes', with the result that each iteration of the
loop in that function outputs a progress update.

The progress update appears at the end of the current line and
overwrites itself with each repetition, producing an output resembling
the following:

    Now eliminating candidates based on [...]: (16/19324)

# Conflicts:
#	Rdutil.cc
#	Rdutil.hh
#	rdfind.cc