Use hash of first and last bytes instead of direct comparison
By default we were using the first/last 64 bytes to quickly
compare if files are the same or not.
This diff gets rid of that and instead hashes the first or
last 1MB of the file and compares the produced hash instead.

While this slows down that initial quick comparison of 64 bytes read to 1MB read + 1MB MD5 hash,
it is quicker than having to later hash large files which is a problem on HDDs with TBs of data.

Future TODO:
- Decide how to handle this via CLI
  1. Add flag to toggle this early partial hash comparison type + flag to select partial hashing algo
  2. Add flag to toggle this early partial hash comparison type but reuse same specified hashing algo
  3. Something else...
- Decide if we want to add flag for number of bytes to partial hash
- Add tests